+++ test/test_league_loader/test_FleaflickerLeagueLoader.py
@@ -1,4 +1,16 @@
 import unittest
+from unittest import mock
+from unittest.mock import Mock
+import copy
+from leeger.enum.MatchupType import MatchupType
+
+from leeger.league_loader import MyFantasyLeagueLeagueLoader
+from leeger.model.league.League import League
+from leeger.model.league.Matchup import Matchup
+from leeger.model.league.Owner import Owner
+from leeger.model.league.Team import Team
+from leeger.model.league.Week import Week
+from leeger.model.league.Year import Year
 
 from leeger.league_loader.FleaflickerLeagueLoader import FleaflickerLeagueLoader
 
@@ -12,3 +24,11 @@ class TestFleaflickerLeagueLoader(unittest.TestCase):
         with self.assertRaises(ValueError) as context:
             FleaflickerLeagueLoader("foo", [])
         self.assertEqual("League ID 'foo' could not be turned into an int.", str(context.exception))
+
+    @mock.patch("fleaflicker.api.LeagueInfoAPIClient.LeagueInfoAPIClient.get_league_standings")
+    def test_loadLeague_happyPath_noOwnerNamesAndAliases(self, mockGetLeaguestandings):
+        mockLeagueStandings2022 = {
+            "divisions": [{"teams": [{"owners": [{"displayName": "Owner 1"}]}]}]
+        }
+
+        mockGetLeaguestandings.side_effect = [mockLeagueStandings2022]