+++ test/test_league_loader/test_MyFantasyLeagueLeagueLoader.py
@@ -1,7 +1,5 @@
 import unittest
-import unittest
 from unittest import mock
-from unittest.mock import Mock
 import copy
 from leeger.enum.MatchupType import MatchupType